namada_vp_env::collection_validation
Trait LazyCollectionExt

pub trait LazyCollectionExt: LazyCollection {
    type Action;
    type SubKeyWithData: Debug;

    // Required methods
    fn read_sub_key_data<ENV>(
        env: &ENV,
        storage_key: &Key,
        sub_key: Self::SubKey,
    ) -> Result<Option<Self::SubKeyWithData>>
       where ENV: for<'a> VpEnv<'a>;
    fn validate_changed_sub_keys(
        keys: Vec<Self::SubKeyWithData>,
    ) -> Result<Vec<Self::Action>>;

    // Provided methods
    fn accumulate<ENV>(
        &self,
        env: &ENV,
        builder: &mut Option<ValidationBuilder<Self::SubKeyWithData>>,
        key_changed: &Key,
    ) -> Result<bool>
       where ENV: for<'a> VpEnv<'a> { ... }
    fn validate(
        builder: ValidationBuilder<Self::SubKeyWithData>,
    ) -> Result<Vec<Self::Action>> { ... }
}

Extensions for LazyCollections for validation.

Required Associated Types

type Action

Actions on the collection determined from changed storage keys by Self::validate

type SubKeyWithData: Debug

Possible sub-keys together with the data read from storage

Required Methods

fn read_sub_key_data<ENV>( env: &ENV, storage_key: &Key, sub_key: Self::SubKey, ) -> Result<Option<Self::SubKeyWithData>>
where ENV: for<'a> VpEnv<'a>,

Try to read and decode the data for each change storage key in prior and posterior state. If there is no value in neither prior or posterior state (which is a possible state when transaction e.g. writes and then deletes one storage key, but it is treated as a no-op as it doesn’t affect result of validation), returns Ok(None).

fn validate_changed_sub_keys( keys: Vec<Self::SubKeyWithData>, ) -> Result<Vec<Self::Action>>

Validate changed sub-keys associated with their data and return back a vector of Self::Actions, if the changes are valid

Provided Methods

fn accumulate<ENV>( &self, env: &ENV, builder: &mut Option<ValidationBuilder<Self::SubKeyWithData>>, key_changed: &Key, ) -> Result<bool>
where ENV: for<'a> VpEnv<'a>,

Accumulate storage changes inside a ValidationBuilder. This is typically done by the validity predicate while looping through the changed keys. If the resulting builder is not None, one must call fn build() on it to get the validation result. This function will return Ok(true) if the storage key is a valid sub-key of this collection, Ok(false) if the storage key doesn’t match the prefix of this collection, or error if the prefix matches this collection, but the key itself is not recognized.

fn validate( builder: ValidationBuilder<Self::SubKeyWithData>, ) -> Result<Vec<Self::Action>>

Execute validation on the validation builder, to be called when accumulate instantiates the builder to Some(_), after all the changes storage keys have been processed.

Object Safety

This trait is not object safe.