namada_vp_env
Trait VpEnv

pub trait VpEnv<'view>
where
    Self: 'view,
{
    type PrefixIter<'iter>
       where Self: 'iter;
    type Pre: StorageRead<PrefixIter<'view> = Self::PrefixIter<'view>>;
    type Post: StorageRead<PrefixIter<'view> = Self::PrefixIter<'view>>;

    // Required methods
    fn pre(&'view self) -> Self::Pre;
    fn post(&'view self) -> Self::Post;
    fn read_temp<T: BorshDeserialize>(&self, key: &Key) -> Result<Option<T>>;
    fn read_bytes_temp(&self, key: &Key) -> Result<Option<Vec<u8>>>;
    fn get_chain_id(&self) -> Result<ChainId>;
    fn get_block_height(&self) -> Result<BlockHeight>;
    fn get_block_header(
        &self,
        height: BlockHeight,
    ) -> Result<Option<BlockHeader>>;
    fn get_block_epoch(&self) -> Result<Epoch>;
    fn get_tx_index(&self) -> Result<TxIndex>;
    fn get_native_token(&self) -> Result<Address>;
    fn get_pred_epochs(&self) -> Result<Epochs>;
    fn get_events(&self, event_type: &EventType) -> Result<Vec<Event>>;
    fn iter_prefix<'iter>(
        &'iter self,
        prefix: &Key,
    ) -> Result<Self::PrefixIter<'iter>>;
    fn eval(&self, vp_code: Hash, input_data: BatchedTxRef<'_>) -> Result<()>;
    fn get_tx_code_hash(&self) -> Result<Option<Hash>>;
    fn charge_gas(&self, used_gas: Gas) -> Result<()>;

    // Provided methods
    fn read_pre<T: BorshDeserialize>(
        &'view self,
        key: &Key,
    ) -> Result<Option<T>> { ... }
    fn read_bytes_pre(&'view self, key: &Key) -> Result<Option<Vec<u8>>> { ... }
    fn read_post<T: BorshDeserialize>(
        &'view self,
        key: &Key,
    ) -> Result<Option<T>> { ... }
    fn read_bytes_post(&'view self, key: &Key) -> Result<Option<Vec<u8>>> { ... }
    fn has_key_pre(&'view self, key: &Key) -> Result<bool> { ... }
    fn has_key_post(&'view self, key: &Key) -> Result<bool> { ... }

}

Validity predicate’s environment is available for native VPs and WASM VPs

Required Associated Types

type PrefixIter<'iter> where Self: 'iter

Storage read prefix iterator

type Pre: StorageRead<PrefixIter<'view> = Self::PrefixIter<'view>>

Type to read storage state before the transaction execution

type Post: StorageRead<PrefixIter<'view> = Self::PrefixIter<'view>>

Type to read storage state after the transaction execution

Required Methods

fn pre(&'view self) -> Self::Pre

Read storage state before the transaction execution

fn post(&'view self) -> Self::Post

Read storage state after the transaction execution

fn read_temp<T: BorshDeserialize>(&self, key: &Key) -> Result<Option<T>>

Storage read temporary state Borsh encoded value (after tx execution). It will try to read from only the write log and then decode it if found.

fn read_bytes_temp(&self, key: &Key) -> Result<Option<Vec<u8>>>

Storage read temporary state raw bytes (after tx execution). It will try to read from only the write log.

fn get_chain_id(&self) -> Result<ChainId>

Getting the chain ID.

fn get_block_height(&self) -> Result<BlockHeight>

Getting the block height. The height is that of the block to which the current transaction is being applied.

fn get_block_header(&self, height: BlockHeight) -> Result<Option<BlockHeader>>

Getting the block header.

fn get_block_epoch(&self) -> Result<Epoch>

Getting the block epoch. The epoch is that of the block to which the current transaction is being applied.

fn get_tx_index(&self) -> Result<TxIndex>

Get the shielded transaction index.

fn get_native_token(&self) -> Result<Address>

Get the address of the native token.

fn get_pred_epochs(&self) -> Result<Epochs>

Given the information about predecessor block epochs

fn get_events(&self, event_type: &EventType) -> Result<Vec<Event>>

Get the events emitted by the current tx.

fn iter_prefix<'iter>( &'iter self, prefix: &Key, ) -> Result<Self::PrefixIter<'iter>>

Storage prefix iterator, ordered by storage keys. It will try to get an iterator from the storage.

fn eval(&self, vp_code: Hash, input_data: BatchedTxRef<'_>) -> Result<()>

Evaluate a validity predicate with given data. The address, changed storage keys and verifiers will have the same values as the input to caller’s validity predicate.

If the execution fails for whatever reason, this will return false. Otherwise returns the result of evaluation.

fn get_tx_code_hash(&self) -> Result<Option<Hash>>

Get a tx hash

fn charge_gas(&self, used_gas: Gas) -> Result<()>

Charge the provided gas for the current vp

Provided Methods

fn read_pre<T: BorshDeserialize>(&'view self, key: &Key) -> Result<Option<T>>

Storage read prior state Borsh encoded value (before tx execution). It will try to read from the storage and decode it if found.

fn read_bytes_pre(&'view self, key: &Key) -> Result<Option<Vec<u8>>>

Storage read prior state raw bytes (before tx execution). It will try to read from the storage.

fn read_post<T: BorshDeserialize>(&'view self, key: &Key) -> Result<Option<T>>

Storage read posterior state Borsh encoded value (after tx execution). It will try to read from the write log first and if no entry found then from the storage and then decode it if found.

fn read_bytes_post(&'view self, key: &Key) -> Result<Option<Vec<u8>>>

Storage read posterior state raw bytes (after tx execution). It will try to read from the write log first and if no entry found then from the storage.

fn has_key_pre(&'view self, key: &Key) -> Result<bool>

Storage has_key in prior state (before tx execution). It will try to read from the storage.

fn has_key_post(&'view self, key: &Key) -> Result<bool>

Storage has_key in posterior state (after tx execution). It will try to check the write log first and if no entry found then the storage.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.